Added error logging to help people cope with the new VM implementation.
Previously, you could get away with
  #foo( bar )
when you really meant
  #foo( "bar" )

The difference is that we only allow 'words' as directive args when they are
a 'helper' word, like 'in' found in #foreach( $a in $b )

So now, it logs it, and the arg is rendered as the context arg for the VM, so
the user visually knows something is wrong.

We could allow it, but wrapping the word into a string lit,
but I think we should nip this now.


PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/trunk@74157 13f79535-47bb-0310-9956-ffa450edef68