Removed all deprecated #set forms, replacing with the normalized form #set().
I had to change two cmp files, escape, to test the escaping of the new #set(), and
include, because the include.vm includes iteself, so the output changed.

Other than that, it appears the #set() perfectly replaces the old #set, including
all whitespaceing issues, for better or worse.
PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/trunk@74167 13f79535-47bb-0310-9956-ffa450edef68