Two tiny tweaks :
1) Resource : altered the 'checking' math to work in terms of seconds rather
than milliseconds, and further, let modificationInterval <= 0 mean 'no check'.
I added the '< 0' as a negative interval is meaningless - once we assign
meaning to it, we can redo.

2) ResourceManager : small tweak in the reloading section to update age and
'checking' counters so that after a reload, everything reset for the next
checking interval.

I think that is all.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/trunk@74260 13f79535-47bb-0310-9956-ffa450edef68