Added a little kludge to work around the infamous "Dreaded <MORE> Parse Problem" (my name, but its a known issue with javacc) which basically boils down to the hosing you get when your input stream ends in a <MORE> token.
This is a general problem that I will fix real soon now as I am sick of
it.

You can't work around it if you have an interpolatable string literal
that is being passed to a directive or a VM :

#foo( "PO#" )

The workaround is

#foo( 'PO#' )

but then you can't do things like

#foo("$fieldname #")

So this fixes it until we can deal with it in javacc, or rework the <MORE>
constructions.

If this works well, will do the same in the general parsing to ensure that no
one runs across it again, anywhere.
PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/trunk@74296 13f79535-47bb-0310-9956-ffa450edef68