Added two things related to the keysAsListed array :
1) protect to ensure that we don't add the same value twice, as this can
result in much pain and suffering when making subsets, especially when
the key is vector-valued - you get a vector that consists of the
normal elements, followed by one or more vectors as elements because
they get added twice.

2) in clearProperty() ensure that we remove the key from keysAsListed. I
think this makes the first safety check redundant, but as this is only
init-time stuff, the miniscule performance hit of 1) is worth it.


PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/trunk@74646 13f79535-47bb-0310-9956-ffa450edef68