ResourceManager.java :
Changed the algorithm for first-time load -
1) We don't note which classloader we are trying - we only log which
loader we found it in (for reference) or note failure if we
don't find it.

2) It's ok if resource.process() throws a ResourceNotFoundException -
that doesn't indicate failure of the resource finding process - only
when done, and no loader was able to provide the resource, is it
considered failure, so we log as such and throw the RNFE.

FileResourceLoader.java :

1) Removed logging - the ResourceManager will log appropriately now.

ResourceLoader.java :

1) Removed the assumption that each loader will have a cache and
modCheckInterval prop associated with it.  (Classpath loader has
neither).  So replaced with 'safe' defaults.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/trunk@74647 13f79535-47bb-0310-9956-ffa450edef68