Medium-term changes to support per-template encoding.  Main motivation is the 'app pack' (or whatever they call it) in Turbine.  Now that there is a customer for it (Turbine) it's good that we do it.
I think an iteration of change might be good, rearchitecting this a little.
Some rework might be in order to change some aspects of stream management for resources.

Resource.java : added encoding as class member, with set/get accessors.
ContentResource.java : added inclusion of the encoding for reading the
intput stream.
ResourceManager.java : need to support encoding in getResource(), as the RM
returns a Resource read to go...


PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/trunk@74874 13f79535-47bb-0310-9956-ffa450edef68