For convertion from Configuration to ExtendedProperties, added a 'shadow' ExtendedProperties to the Configuration class.  Why?
in o.a.v.apps.Velocity, there is a method setConfiguration() that we want to
deprecate.  The only way to do this would be to take the Configuration,
copy all data out and into an ExtendedProperties, and then init
the Runtime with that.  Now, that's not easy w/o modifying the EP to
support that (I think), so the easiest was to shadow like this.

I don't think anyone except turbine uses this anyway, and turbine will change
too :)
PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/trunk@74944 13f79535-47bb-0310-9956-ffa450edef68