This is a test to see if we can move to ant 1.4 w/o everyone squawking or having problems. The major motivator is quieting gump while we work things out with the ant developers. (it would break for users too, possibly.)
The changes in testcases.xml are to use the 1.3 classloader approach to
things will work with Gump (and any user that has the vel.jar in their
classpath.)  I am working with the ant devs to address this.  These
changes are 1.4 specific, and sadly, 1.3 doesn't like attributes that it
doesn't know about, so this now won't work with ant 1.3.

Removing xerces and xalan as they get in the way with ant 1.4.

So w/o the jars and the changes to the .xml, under ant 1.4 all is well
w/ and w/o the vel.jar in the environment CLASSPATH.  This might not make
it in in time for the morning gump run, but the messages should stop after
this (unless something else breaks... that's why I want to quiet these...)

If anyone has any problems with this, we should/must discuss - I am very
averse to creating hardship with this and therefore not married to it.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/trunk@75332 13f79535-47bb-0310-9956-ffa450edef68