Fix for bug # 5920.
I basically rewrote it, and would really appreciate someone who uses this
to look over the changes.  I think it's right, but I added some
semantics which were undefined, namely it returns the input if the
input string is null, the count is 0, or the EOL is null.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/trunk@75588 13f79535-47bb-0310-9956-ffa450edef68