Added JavaDoc to process() spurred by investigatin of Ilkka's "Less messages from ContentResource" patch.
The JavaDoc I added brings up an interesting dichotomy between the
Template and ContentResource classes -- the former currently always
throws a ResourceNotFoundException when it can't find a template,
while with Ilkka's patch the latter will be returning false instead.
It seems to be that these should act in a consistent manner, so
perhaps I'll simply have ContentResource re-throw any
ResourceNotFoundException it encounters in process() (rather than
returning false).

Though this makes things consistent and seems to use
ResourceNotFoundException as intended, it still leaves the API open to
too much interpretation.  More JavaDoc on the process() method which
suggests when it is appropriate to return false would be very help.

Please chime in on this one guys.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/trunk@75769 13f79535-47bb-0310-9956-ffa450edef68