static Introspector object? Now that scares me.
And makes me wonder if ever anyone has tried to run two different
VelocityEngine instances with different Introspectors in the same
VM. Make it a protected instance member. private and getters/setters
might even be better.

setLog() initialized the introspector object??? That can't be
right. The init code in the runtimeInstance makes sure that init() is
called last (after setLog() if the class implements UberspectLoggable)
so we can move this into init() where it belongs.

Do some javadoc fixup. 




git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/core/trunk@292978 13f79535-47bb-0310-9956-ffa450edef68