Remove String(String) C'tor.
Can anyone tell me whether this whole code segment is actually
correct? Because a method returning 'null' and not having a void
return type might cause this code to return 'null' to the
caller. According to the comment on top of this block, this might
actually be a bad idea. Maybe we can just remove the whole
method.getReturnType() check and return unconditionally "" when the
result is null?



git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/core/trunk@292980 13f79535-47bb-0310-9956-ffa450edef68