Remove libraries from Subversion. The ant build now fetches all needed libraries directly from the central repository at www.ibiblio.org.
You need this only if you actually fetched the source code from the
web and want to compile it for yourself. As for the download a web
connection is needed anyway, there is no point in fetching the jars
from the Apache Subversion repo.

This also reduces the size of the actual download from the Subversion
by ~3.5 MB.



git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/core/trunk@329218 13f79535-47bb-0310-9956-ffa450edef68