This is an example patch to switch to the ivy dependency resolver for building Velocity.
Advantage: We would get all the dependency downloading and resolving
features that maven has, in a nice rounded package. If might be
possible to do that with Maven Wagon too, if there were any docs on
how to use wagon inside ant (laughs).

Disadvantage: Increases the size of the source distribution by ~350k
because we must include the ivy jar.



git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/engine/trunk@383388 13f79535-47bb-0310-9956-ffa450edef68