Add a wrapper to be able to prefix and control the output of log messages on a finer grained base than just global logging. VelocityMacroFactory did this internally for a long time and having that in a separate class usable from different place is probably a good thing.
Needs unfortunately (Log is not an interface) some inheritance magic
to be able to fully replace the Log object inside a class and still
note when the chute object inside the "parent log" changes. 

This also needs some minor changes to Log itself (never access the
member variable directly but only through the getter, so we can
override it in the inheriting class).




git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/engine/trunk@462516 13f79535-47bb-0310-9956-ffa450edef68