Factor out the Introspector cache. Add a cache listener interface and also an interface definition for the Cache itself. Move all functionality into the cache class, leaving Introspector and IntrospectorBase as very lean classes. This shouldn't actually change too much of our public interface, but a lot of the protected methods, but they were never part of our API anyway.
This change allows Velocity users to register listeners to the Cache and react
on events. There is now also a clear definition on what the Cache must implement
to be pluggable (what is still missing is actual lookup of the implementation class
from properties. Shouldn't be too hard to do.

First steps in the direction of fixing VELTOOLS-66 once and for all.

git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/engine/trunk@476951 13f79535-47bb-0310-9956-ffa450edef68