Support Java 16 records in property lookup
Java 16 records (JEP 395) are a perfect fit for Velocity variables: they
are lightweight and desgined to contain immutable data. However, records
don’t follow the JavaBean convention: the getter of the field `foo` is
`foo()`, not `getFoo()`. Add support for this naming convention.

While this naming convention is mostly useful in projects using Java 16
records, the patch itself doesn’t rely on records, so it doesn’t break
Java 1.8 compatibility.